Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3:
> > > 
> > > Note this leads to errors on chromebooks:
> > > https://build.chromium.org/p/client.v8.ports/builders/V8%20Arm/builds/2228
> > > https://build.chromium.org/p/client.v8.ports/builders/V8%20Arm%20-%20debug/builds/2172
> > 
> > Uh oh.. I'll take a look at this ._.
> 
> Looks like there's an issue with the runtime implementation on ARM relating to +/-Infinity values.
> 
> I can take a look at that locally, but feel free to roll out this and the followup refactoring if needed. I should have a fix in a few hours.

It's going to take a bit longer than a few hours, as cpu emulation is making this take forever. I'm seeing if we have a proper ARM device I can use to take a look at this.

In the mean time, if this is a problem, I suggest a rollout (or an update to test262.status as this should be fixed within the week)

Patch-set: 3