Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Do you have some kind of benchmarks for this? In my experience there are significant performance penalties when using C++ vs. CSA, so please just use that if performance is relevant.

This is brand new, and the remaining features from the proposal haven't landed upstream yet.

This will most definitely be slower in virtually all cases, due to the calls from JS to C++, and back into JS.

But, there's no users of this language feature yet, and the full set of this language feature hasn't landed yet, so it's probably OK to be slow to begin with.

Patch-set: 2
Reviewer: Gerrit User 1162379 <1162379@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>