[Compiler] Remove CompileBaseline and three tier pipeline.
Since we no longer support the ignition-staging configuration any longer,
we can retire the three tier pipeline and the CompileBaseline
functionallity.

We still need support for JSFunction self healing due to liveedit
(which for --no-turbo might end up replacing a forced Ignition function
with a FCG function) - we can remove this once we remove --no-turbo support.

BUG=v8:4280

Change-Id: I5482abd17785324654e022affd6bdb555b19b181