Update patch set 1
Patch Set 1:

So far, this seems to be compatible with SpiderMonkey.

Babel seems to complain about `var let = [1, 2]` (which I wanted to use to test `let [a] = [b]` in the conditional expression), so Babel is (I believe) wrong about that in particular. But as mentioned, v8 allows it. I removed the test from test-parsing.cc because it's not exactly destructuring assignment, though.

Patch-set: 1
Reviewer: Gerrit User 1162379 <1162379@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>