Update patch set 36
Patch Set 36:

We should discuss a plan B design on the whiteboard to understand how the MarkCompact would look like if we do not use templates. We should do the cleanest solution. If it is templates, fine. If not, we should do something else (whatever it is).

Patch-set: 36
Reviewer: Gerrit User 1119605 <1119605@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>