Update patch set 1
Patch Set 1:

High level comment - I find the new approach for collecting feedback in the non-smi/number cases to be much more confusing than the old approach. Does it generate much better code? Could we just use the same approach as previous (maybe with some tweaks so that we don't have to recheck something we already checked in the fastpath) ?

Patch-set: 1