Update patch set 2
Patch Set 2:

(1 comment)

This would be a candidate for squashing into a generic "Remove Isolate from ParseInfo" CL... but I can see now that the changes are slightly bigger than I expected, so I'm slowly changing my opinion about whether squashing is the right thing.

vogelheim@, wdyt? The ultimate goal is to be able to create and run parsing tasks on a background thread, and for that, we must be able to create ParseInfos without Isolate. As we can already parse on a background thread, it would be intuitive that ParseInfo doesn't have Isolate, only CompilationInfo does. A bunch of CLs implement that. Squash or no squash?

Patch-set: 2
Reviewer: Gerrit User 1115961 <1115961@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>