Update patch set 2
Patch Set 2:

(1 comment)

I'm uncomfortable with this logic. As I understand it, this effectively introduces a two-phase initialization, where some members are initilaized in the constructor, and some when an Isolate* is set. Currently, every ParseInfo cosntructor will have an Isolate* in its constructor, so this should work, but there's no guarantee (or safeguards) that this will still work when that changes.

What I'd like to see is:
- If we always require an Isolate*, then it should go into the constructor.
- If we always require an Isolate*, but only later in ParseInfo's lifecycle, then we should have some assertions for those later-in-the-lifecycle methods that ascertain that an Isolate* has been set.
- If we don't require an Isolate*, then we should have some assurance (and tests) that the code will actually work with those Isolate-dependent fields not being set.

Patch-set: 2
Reviewer: Gerrit User 1117099 <1117099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>