Update patch set 4
Patch Set 4:

Is there a design doc for the type profile collection? I'm a bit worried that there are going to be different bytecodes for every different type collection point. Also, did we consider using the approach Yang mentioned in the doc of having the bytecode offset as an index and letting the debugger hijack store bytecodes to profile types? 

Let me know if you want to chat about this in person, happy to jump on GVC or I'll be in Munich from Wednesday next week.

Patch-set: 4
Reviewer: Gerrit User 1002763 <1002763@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>