[Interpreter] Inline fast paths and call builtin for others in compare handlers.
Compare bytecode handlers does several checks to collect feedback and
repeat these checks when actually performing the operation. This cl, changes
it perform the actual comparison for smi/number along with collecting feedback.
For the other cases, the handler collects the feedback and calls the correct
builtin to perform the operation. This reduces the size of the bytecode handlers.

BUG=v8:4280

Change-Id: If0954b68dd232461e08e94a90ccc17604c235b27