Update patch set 3
Patch Set 3:

I uploaded an alternative fix which doesn't create a FunctionState in ParseArrowFunctionLiteral, but but... Toon, does that look better to you? I don't find it better than the orig fix. It avoids one RewriteDestructuringAssignments call though.

Btw, there are also two FunctionStates for non-arrow functions when doing DoParseFunction: one is in DoParseFunction and the other in ParseFunction (the latter one, the one which is called via DoParseFunction -> ParseFunctionLiteral -> ParseFunction). I think the outer one is supposed to represent the outer scope outside the function.

Patch-set: 3