Update patch set 5
Patch Set 5:

Routing offline discussion w/ adamk@ here:

We agree that the first fix (patch set 1) is better. Here's why:

In DoParseFunction (and functions called by it), two FunctionStates are created, both for arrow and non-arrow functions, so that part is fine.

When parsing an arrow function, the destructuring assignments in the params go to the "parent" FunctionState. This is true for eager and lazy cases. In the lazy case we *do* know that we're parsing an arrow function, so we could force all of them into the same FunctionState (which is the alternative fix, patch set 5). But not forcing it is hmm... leaner? more consistent? less of a hassle?

Patch-set: 5
Reviewer: Gerrit User 1115961 <1115961@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>