Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+1
> 
> LGTM, and agree with Jacob that a regression test would be nice

I also agree that a regression test would be nice. The problem is, as I discussed with Benedikt, that crash most likely only happens in Crankshaft, and by the nature of the problem a regression test would likely be flaky (by flaky I mean that some changes in the code generation could make the test useless). To sum up, I don't know how I could write a reliable test for this bug. I tested the CL on my phone though in the no-turbofan configuration on a webpage where I was able to reproduce the crash, and it fixed the problem.

Patch-set: 2
Reviewer: Gerrit User 1141154 <1141154@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0