[wasm] --trace-turbo in CreatePipelineStatistics needs a parse_info
A recent CL enabled pipeline statistics for WebAssembly. This caused a
problem with the --trace-turbo flag because in the pipeline statistics
code --trace-turbo wanted to access the parse_info, which is not
available for WebAssembly. With this CL I guard the trace-turbo code
behind a parse_info check to avoid this problem.

R=clemensh@chromium.org

Change-Id: I9d628c7dec5b456e0ff9178ad989c41ac1e0237e