Update patch set 6
Patch Set 6:

FYI, CL to rip out the variables in gyp:
https://chromium-review.googlesource.com/c/461145/2

Through I believe it doesn't fix the cross-compilation issue above. Ivica, could you please file a bug on v8 tracker for Aleksey and/or Yang with all the details to get more attention to it? This reply history here might fall through the cracks.

Patch-set: 6