Update patch set 4
Patch Set 4:

(2 comments)

Nice!

I had a look at the spec and I'm confused if we need to do more, because SingleLineDelimitedComment is also removed from the grammar in the case of modules.  However, MultilineComment is changed in a way that may make up for that?!  To me this actually looks like the two parts of the spec got out of sync.

It also looks to me like the following should be illegal in classic scripts but we currently accept it (note that it's a single line):
/**/-->

Patch-set: 4
Reviewer: Gerrit User 1119597 <1119597@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>