Update patch set 1
Patch Set 1:

(4 comments)

> When I was last looking at the deopt_count, it seemed really broken -- I believe the reason was the counting of lazy deopts after eager deopt. As far as I remember, the optimized code indeed gets unlinked from the JS functions, but the deoptimizer still knows the function, so I think it would still bump the deopt-count. Perhaps best is to construct some examples of this, add some printfs and see what happens.

Sorted, by keeping an "already deopted" bit on Code objects -- this also helps with recursive calls into a soft deopt not counting as a deopt. Interestingly, what was happening was that the deopt count was going up to 2, one for the eager and one for the lazy.

> > Good question, right now the deopt count wraps around. Maybe it should saturate instead?
> 
> Yeah, that sounds better.

Done.

> When we unlink, we already walk over all JS functions, so we should be able to get to the SFI from there.

Good point, thanks. Done.

Patch-set: 1
Label: Commit-Queue=0