Update patch set 1
Patch Set 1:

> Patch Set 1: Commit-Queue+2
> 
> > Patch Set 1: Code-Review+1
> > 
> > Nice cleanup, thanks! I suppose this is left-over from the time where we either still called setters even though the property was read-only, or from the time when we called such setters through the prototype chain with incompatible receivers... Both are long fixed, so nice to get rid of this as well.
> 
> Thanks for the review. Do you happen to have any insight into why parts of the bootstrapper call SetPrototype and others SetInstancePrototype? Should they all call the latter (for speed)? Or the former (for guaranteed correctness)? Given that all this stuff should live in the snapshot, I wouldn't think speed would be a factor.

There's probably no real reason. It gives some guarantee that we don't accidentally set a non-instance-prototype in the bootstrapper. It would be even nicer if SetInstancePrototype would have JSReceiver is type for value though.

I'd be fine with just always calling SetPrototype and inlining SetInstancePrototype into it. I agree that performance isn't important. And the added guarantee that it is an instance prototype probably doesn't outweigh the simplicity.

Now that I look at it, it seems like SetPrototype can unnecessarily keep memory alive? If we first set a non-instance prototype, and then set it back to an instance prototype, the constructor field of the function will still point to the old non-instance prototype. Wanna fix this along the way as well? :-) (If I read the code correctly at least.)

Patch-set: 1
Reviewer: Gerrit User 1184114 <1184114@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>