[esnext] implement spec change to TaggedTemplate callsite caching
Proposal: https://github.com/tc39/ecma262/pull/890

Description of implementation:

1. AST nodes are added to represent template literals, and tag invocations, replacing the parse-time desugaring and runtime calls

2. Generate bytecode for the string concatenation in
BytecodeGenerator::VisitTemplateLiteral(), using a single assignment op
feedback slot (which is safe, because added items are always strings, except
.

3.Generate similar bytecode to Call nodes in
BytecodeGenerator::VisitTaggedTemplate(), with the exception that the
first parameter becomes a new LoadTemplateObject bytecode.

4. LoadTemplateObject grabs the SharedFunctionInfo from the closure register, loads FixedArray element at tagged template index, and calls runtime if it's not a JSArray (which creates the object and stores it in the cache).

5. TurboFan turns LoadTemplateObject into a HeapConstant (if the template cache entry exists), otherwise performs the slower conditional runtime call.

---

Difficulties:

The above approach generally works well, but runs into problems when ComplationCache is used, as the same SharedFunctionInfo can be used in multiple contexts (in which case, cached entries added in one context will be invalid from a different context). This causes stress-opt failures.

A proposed solution to this problem is to assume that the SFI is only used in
a single Context at a time (which I believe may be sufficient for Chromium
due to render processes having their own Isolate & cache, but I'm not
positive!), and keep a pair of items in the cache for each entry (a WeakCell holding the native context, or undefined). If the WeakCell value is not the current native context (in the LoadTemplateObject bytecode), the cache entry is invalid and must be re-created (to prevent leaking a cross-context Array.prototype).

This is of course more complicated, more costly in terms of memory, and it's possible there's a better way to do it.

BUG=v8:3230, v8:2891
R=

Change-Id: I98dd5e2d4ddd8a1d85522f2314a52c3a04d02536