Update patch set 9
Patch Set 9:

(1 comment)

> Patch Set 9:
> 
> I'm not happy with having three different attributes (initialization_flag, binding_needs_init, ForceHoleInitialization) whose names/values suggest they all determine the same thing, except they don't.  Maybe it's enough to choose the names more carefully, but I'm hoping we can somehow simplify this a bit.  Adam, would you mind chatting about it?

I agree that this situation is not ideal, happy to discuss tomorrow.

I just spent a few minutes trying to make this better, but I wasn't satisfied with the result. The overall issue is that there are many different aspects of Variables (VariableKind, VariableMode, InitializationFlag), some of which are implied by each other, some of which can be optimized away. Capturing all of that in an easy-to-understand way is hard.

Patch-set: 9
Label: Commit-Queue=0