[wasm] Make CompilationHelper own the WasmModule
Another fix for more explicit management of ownership. The
CompilationHelper now always owns the WasmModule, and transfers
ownership to the generated WasmModuleWrapper (a Managed<WasmModule>)
once that object is created. Since the stored uniqe_ptr cannot be
accessed any more after this transfer, the creation of the
WasmModuleWrapper is delayed until it is really needed (step 5 in async
compilation).

R=ahaas@chromium.org

Change-Id: I22dea2e14a364ddf76751d97bd0d736a4c0ceff4