[wasm] Change the error_pc in the decoder to error_offset
The error_pc was only used to calculated the relative error offset. By
switching to an error_offset directly will allow us later to give the
decoder a base offset eventually so that we can get correct error
positions even when the decoder is executed on multiple memory chunks,
which will happen with streaming compilation. With this change I also
had to provide "kind of" reasonable error position in the
StreamingDecoder.

R=clemensh@chromium.org

Change-Id: I736fa082c51c64334d23771061acf97e2c47778e