[wasm] Streaming processor
This CL introduces the StreamingProcessor class, which is called by the
StreamingDecoder whenever a section or function has been received
completely. My plan is that eventually the AsyncCompileJob implements
this interface to start compilation jobs. Error handling is also taken
over by the StreamingProcessor.

I introduce the StreamingProcessor to decouple the decoding and
buffering of the stream from the actual compilation. This makes it
easier to test each part separately.

R=clemensh@chromium.org, mtrofin@chromium.org

Change-Id: Ie9d987c2400382f2d7bbef6d11ed74610d8ab175