[runtime] avoid trim/grow loop when adding and removing one element
We currently grow the backing store to (old_capacity*1.5)+16 if we exceed capacity, 
but shrink the capacity to the current length when 2*length <= capacity.
For short arrays (up to length 32), this can lead to a copy on every operation when using push/pop or push/shift.

Example:
Array of length 32, capacity 32
push
Array grown to length 33, capacity 32*1.5+16 = 64
pop
Array trimmed to length 32, capacity 33 because 2*32 <= 64
...

This CL restricts trimming to cases where we free at least 32 elements plus the new length of the array.
This prevents the described loop and also prevents trimming on short backing stores: the backing store is never trimmed if its size is at most 32.
It does, however, not prevent shrinking to very small backing stores.

Bug:
Change-Id: I9dd13e5e2550c7ac819294c8e29f04c8855e02a4